chtml/html/testdata/webkit/svg.dat

#data
<svg><tr><td><title><tr>
#errors
#document
| <svg svg>
|   <svg tr>
|     <svg td>
|       <svg title>
|         <tr>

#data
<svg><thead><title><tbody>
#errors
#document
| <svg svg>
|   <svg thead>
|     <svg title>
|       <tbody>

#data
<svg><tfoot><title><tbody>
#errors
#document
| <svg svg>
|   <svg tfoot>
|     <svg title>
|       <tbody>

#data
<svg><tbody><title><tfoot>
#errors
#document
| <svg svg>
|   <svg tbody>
|     <svg title>
|       <tfoot>

#data
<svg><tbody><title></table>
#errors
#document
| <svg svg>
|   <svg tbody>
|     <svg title>

#data
<svg><thead><title></table>
#errors
#document
| <svg svg>
|   <svg thead>
|     <svg title>

#data
<svg><tfoot><title></table>
#errors
#document
| <svg svg>
|   <svg tfoot>
|     <svg title>
